backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
